pyproject.toml
[tool.poetry]
name = "rasenmaeher_api"
version = "1.6.4"
description = "python-rasenmaeher-api"
authors = [
    "Aciid <703382+Aciid@users.noreply.github.com>",
    "Eero 'rambo' af Heurlin <eero.afheurlin@iki.fi>",
]
homepage = "https://github.com/pvarki/python-rasenmaeher-api"
repository = "https://github.com/pvarki/python-rasenmaeher-api"
license = "MIT"
readme = "README.rst"

[tool.poetry.scripts]
rasenmaeher_api = "rasenmaeher_api.console:rasenmaeher_api_cli"


[tool.black]
line-length = 120
target-version = ['py38']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  |  __pycache__
)
'''

[tool.mypy]
strict=true

[[tool.mypy.overrides]]
module="tests"
disallow_untyped_decorators=false

[tool.pytest.ini_options]
junit_family="xunit2"
addopts="--cov=rasenmaeher_api --cov-fail-under=65 --cov-branch"
asyncio_mode="strict"
asyncio_default_fixture_loop_scope = "session"

[tool.pylint.MASTER]
extension-pkg-allow-list = "pydantic"

[tool.pylint.format]
max-line-length = 120

[tool.pylint.design]
max-parents = 10

[tool.pylint.messages_control]
disable=["fixme", "W1202", "C0209"]

[tool.pylint.similarities]
min-similarity-lines = 8
ignore-imports = true

[tool.coverage.run]
omit = ["tests/*"]
branch = true

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
libadvian = "^1.6"
click = "^8.0"
fastapi = ">=0.89,<1.0"  # caret behaviour on 0.x is to lock to 0.x.*
# FIXME: Migrate to v2, see https://docs.pydantic.dev/2.3/migration/#basesettings-has-moved-to-pydantic-settings
pydantic= ">=1.10,<2.0"
pydantic-collections = ">=0.5,<1.0" # caret behaviour on 0.x is to lock to 0.x.*
requests = "^2.31"
multikeyjwt = "^1.0"
uvicorn = {version = "^0.20", extras = ["standard"]}
gunicorn = "^20.1"
pyopenssl = "^23.1"
libpvarki = { git="https://github.com/pvarki/python-libpvarki.git", tag="1.9.0"}
openapi-readme = "^0.2"
python-multipart = "^0.0.6"
aiohttp = ">=3.11.10,<4.0"
pyjwt = ">=2.10.1"
aiodns = "^3.0"
brotli = "^1.0"
cchardet = { version="^2.1", python="<=3.10"}
filelock = "^3.12"
python-keycloak = "^4.2.0"
sqlmodel = ">=0.0.22,<1.0"
psycopg2 = "^2.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
coverage = "^7.3"
pytest-cov = "^4.1"
pylint = "^3.3"
black = "^25.1"
bandit = "^1.8"
mypy = "^1.15"
pre-commit = "^3.3"
pytest-asyncio = ">=0.23,<1.0" # caret behaviour on 0.x is to lock to 0.x.*
bump2version = "^1.0"
detect-secrets = "^1.2"
httpx = ">=0.23,<1.0" # caret behaviour on 0.x is to lock to 0.x.*
types-requests = "^2.31"
types-pyopenssl = "^23.1"
async-asgi-testclient = "^1.4"
pytest-docker = "^2.0"
flaky = "^3.8"


[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
